(FM-3940) Fix specs - Puppet 4.x validation errors
The validation error message for missing parameters for defined
types has changed from "Must pass ___ to ___" to "expects a value
for parameter '___'". The specs must be evaluated based on the
version of Puppet they are running under.

Add specs for config to test missing admin_user/admin_pass to
verify the values are validated properly.