Add windows 2008r2 stuff
The powershell that ships with 2008r2 is version 2 which needs the
Import-Module before Add-WindowsFeature will work, and uses
Add-WindowsFeature instead of Install-WindowsFeature. This should also
continue to work on windows 2012.

The sqlserver_instances fact returns nil when no instances are present
but this was not handled by the provider, so this patch was also needed
on puppet-agent 1.3.2. I'm not sure why this wasn't caught before.