(FM-5041) Install Module from Forge
To enable installing the module from the module staging forge a PE license
must be present on the master. The test helpers have been updated to add a
method for installing a valid license on a master. The pre-suite for
install the "puppetlabs-sqlserver" module along with dependencies was
updated to install the modules onto the master then use plug-in sync
to install the modules onto agents.

[skip ci]