(MODULES-3355) Fix Role Name Collisions
If sqlserver::role is specified more than one for the same role
non-unique titles will be generated for the sqlserver_tsql that
will handle the operations. This is likely to happen if setting
up heira configs like this:

'sqlserver::role':
  ‘User_1’:
    role: db_owner
    database: FooBar
    instance: MSSQLSERVER
    members:
      - ‘User_1’
    type: 'DATABASE'
  ‘User_2’:
    role: db_owner
    database: Wakka
    instance: MSSQLSERVER
    members:
      - ‘User_2’
    type: 'DATABASE'

This fix introduces the database name into the title created for
the sqlserver_tsql statements so that it is unique.