(MODULES-3355) Fix acceptance tests for Sqlserver role
Commit 2f18a684cf8ea49f0aa4c943197762fdf53132d3 fixes
sqlserver::role to allow multiple role declarations for
the same role name. This introduced a bug by not changing
the sqlserver_tsql title in the require paramter for
sqlserver::role::permissions. This only occurs when
permissions are specified.

This fixes acceptance tests by ensuring the require statement
for the role permissions is the same as the tile for
the tsql statement