(PE-16132) Add Windows Based Authentication for sqlserver::config
Previously to use classes such as sqlserver::login, it would use the
sqlserver::config resource which would contain the credentials to conect to the
database.  However this resource could only authenticate using SQL Server based
authenctiation.  This means that databases that only used Windows based
authentication could not be managed.

This commit modifies the sqlserver::config class with an additional property
called login_type which can be either SQL_LOGIN or WINDOWS_LOGIN, with a default
 of SQL_LOGIN.  The login type also determines the validity of the admin_user and
 admin_pass properties.

The config resource properties are consumed by the get_connection_string method
in sql_connection.  The login_type property then determines whether
'Integrated Security=SSPI' is appended the connection string for connecting to
the database.

Additional the spec, define and acceptance tests are modified for the new
properties and default values.