(FM-5387) Remove simplecov gem
FM-5387 simplified the windows CI pipeline complexity however an issue was found
in ths SQL Server module.  The simplecov gem is used to generate code coverage
metrics however the results  aren't actually used anywhere.  Also this gem
depends on json with native extensions which is causing CI failures due to lack
of devkit.  This commit removes the simplecov gem and removes the references in
the spec_helper.