(MODULES-3427) Allow multiple accounts to be used for the SYSADMIN role
Previously there was a bug in the way the module constructed the command line
to install SQL server with multiple SA role accounts.  Each additional account
should be a separate command line parameter instead of combining all accounts
into a single parameter.  This commit changes how the command line arguments are
constructed so that each admin account has its own element in the command line
argument array.  This commit also modifies the spec tests to ensure this
behavior is used.