(MODULES-3256) Fix create/delete of Windows based logins
Previously the module would fail to create and delete windows based logins due
to the TSQL scripts querying the `SQL_LOGINS` table.  This table only lists SQL
logins, not Windows logins.  Instead the `SERVER_PRINCIPALS` table is used to
determine login existance and correctness.

This commit also modifies how Windows Group logins are disabled.  Groups can
not be disabled, they can only be denied to connect to the database.  This
requires the use the GRAND/REVOKE server permissions table.  This commit detects
the state of the login using the `server_permissions` table.