(MODULES-2323) Fix deleting an SQL login
Previously when attempting to remove a login from an MS SQL Server, a TSQL error
was thrown.  This is due to the `GO` statement in the template.  `GO` is not a
TSQL statement, it is actually a command interpretted by a client, such as
sqlcmd.exe.  This commit changes the template to use only TSQL statements and
uses the `DROP LOGIN` command as the `sp_droplogin` stored procedure has been
deprecated.

This commit re-instates the login deletion after each login acceptance test.

Ref: https://msdn.microsoft.com/en-us/library/ms189767.aspx