(MODULES-2554) Fix remove database
Previously when attempting to remove a database from an MS SQL Server, a TSQL
error was thrown.  This is due to the `GO` statement in the template.  `GO` is
not a TSQL statement, it is actually a command interpretted by a TSQL client,
such as sqlcmd.exe.  This commit changes the database delete template to use
only TSQL statements and terminate with a semi-colon.  This change also
reinstates the database deletion in the login_spec and database_spec acceptance
tests.