(MODULES-3256) Fix creating a login resource with additional attributes
Previously when creating a login in a SQL Server it would take to puppet runs
to complete the configuration.  This is due to the ordering of actions in the
`create/login.sql.rb` template.  It would only modify a login if it already
existed.  This commit changes the logic of the script to create the login if it
does not exist and then modify the login with the correct attributes.