(MODULES-3752) Fix modifying server_roles for an existing WINDOWS_LOGIN
Previously the SQL template for modifying a login and the associated acceptance
tests were using the IS_SRVROLEMEMBER function to determine membership.  However
this function has many caveats e.g. failing for domain principals if a domain
controller is not contactable, and always return false for a disabled Windows
user login.  This commit changes the check to use the `sys.server_role_members`
table which holds the underlying membership information.