[PE-17491] Do not fail on install when a restart exit code is returned
When exit code 3010 is returned from setup.exe upon install the puppet run failed even though the install succeeded, the 3010 exit code just signifies that some changes are pending and a restart is required. This change allows the puppet run to succeed when the install returns 3010 exit code and simply raises a warning.