(MODULES-4257) Modify instance and features for SQL Server 2016
Previously the sqlserver_instance and sqlserver_feature acceptance tests were
failing.  This was mainly due to SQL Server 2016 not having a feature for the
SQL Management Studio (ADV_SSMS and SSMS).  However it was later found that the
super features called Tools and SQL were not providing an adequate abstraction
e.g. they were not rolled up in the Facter facts.  This commit adds support for
SQL Server 2016 by the following:
- Deprecating the use of 'Tools' in the sqlserver_features and 'SQL' in the
  sqlserver_instance custom types.  These super features will continue to
  function however use of them will generate a deprecation warning
- The acceptance tests for the super features was removed
- Most tests which referred to ADV_SSMS and SSMS were removed.  Where not
  possible, they were guarded to not be tested on SQL Server 2016.
- An installation media version detection method was added as Dot Net 3.5 is
  not required for SQL Server 2016 installation.  This is only applicable for
  SQL Server 2012 and 2014.
- Refactored the instances method for sqlserver_features to support multiple
  server editions on a server but until MODULES-5060 is resolved, only the first
  set is returned.