(MODULES-5092) Failures during .Net 3 installation should be obvious
Previously if dot Net 3 experienced an error during installation, no error
would be logged and instead the SQL Setup would then fail but not inform Puppet
as to why.  This commit changes the dot Net 3 installation method to instead
log the output from the installation process and fail the puppet run, with log
information, if the installation fails.