(FM-5389) Fix sql_features when installing SNAC_SDK
Previously the sql_features was not idempotent when installing the SQL Native
Client SDK (SNAC_SDK).  This Due to the SQL Native Client not being unique
across SQL Server versions (e.g. SQL 2016 (v13) installs Native Client with a
version that matches for SQL 2012 (v11)).  Due to MODULES-5060 it is noted that
the SQL Server Module can not be used to manage different SQL Server versions
on the same host.  Knowing this it is safe to collate all of the SQL Server
shared features, regardless of version, and use that as the list of installed
features and is safely idempotent.

This commit also adds a README section to known issues about this issue.