(MODULES-2543) Purge members from SQL Server Role
Previously the purge_members parameter of the sqlserver::role resource was not
taking effect.  This commit re-instates the acceptance test for this behaviour
and changes the sql query used to detect members which are no longer required.
It appears the table variable usage was returning zero results for the row
count.  Instead the detection is changed to just using the SELECT query, instead
of an INSERT INTO, which does not require any row count calculation.

This commit also modifies the role deletion as SQL Server requires any members
to be removed prior to a role being deleted.