(FM-5021) Escaping identifiers when creating a database
Previously the name of file and log files were not escaped properly which caused
SQL Server to interpret incorrectly.  The dashes in the name were being seen as
a minus operator.  This commit changes the name of the Log and File files to be
seen as strings instead of identifiers.  This commit also updates the spec tests
to expect these names to be escaped.