(MODULES-4915) Remove forced TCP connection for SQL management
Previously the module tried to connect to the SQL instance using the name
localhost.  This had the side effect of always using the TCP based connection.
If a customer used a non-default Port (common practice) then the sql_connection
would not connect.  This commit instead changes the Datasource to use the '.'
moniker which denotes the local machine.  However this instructs the OLE DB
provider to use the enabled protocols enabled for the native client instead of
only using TCP.  This means it will prefer to use the Shared Memory protocol if
enabled.  Also the Shared Memory protocol is required to be enabled for the SQL
Server Agent process on SQL Server 2016.