(MODULES-5566) Discover instances via registry
Without this commit, the logic to discover instances and features
relies on WMI calls and fails to discover instances unless the
SQL Engine is installed; this meant that Puppet would never
discover instances with only analytics services or reporting
services installed, causing each puppet run to attempt to
reinstall them, even though they exist.

This commit refactors the logic for discovering SQL instances to
use registry instead of WMI. In order to do so, we use the Puppet
Util Windows Registry module and ensure that SQL instances with
valid names outside the current code page can be discovered.

This ensures that we are able to discover SQL instances regardless
of which features they have installed.

This commit also removes code supporting calls to WMI to discover
SQL instances.