(MAINT) Correct supported versions & platforms
This commit removes the specification of minimum required Puppet
Enterprise version from the Setup Requirements section of the
readme. It does so both because this information already exists
in the metadata.json and because the forge prominently displays
compatible versions.

This commit also adds Windows Server 2016 as a supported option
in the same section of the readme since this is also specified
in the metadata.json file.