(MAINT) Remove extraneous logic from spec
This commit removes logic from the features file in the acceptance
tests suite which existed only to support Puppet 3. This logic is
no longer needed as Puppet 3 is no longer supported. Additionally,
the way the logic was written caused the tests designed for Puppet
3 to run against Puppet 5.

This commit simplifies the maintenance and execution of the suite
of acceptance tests.