(MODULES-6022) Fix Instance Unnecessary Reinstall
Prior to this commit, applying a manifest that ensures a SQL
instance is installed with any features will always attempt to
uninstall and reinstall those features on subsequent applies.
This was due to a bug in the code returning discovered features
nested inside of a needless array. This commit rewrites the logic
for retrieving the instance features to concatenate them into an
array instead of mapping them, preventing the array-of-arrays
problem entirely.