(MODULES-6022) Ensure instance idempotency in tests
Prior to this commit, the acceptance tests for sqlserver instances
do not verify idempotencies for the applied manifests. This allowed
us to ship revisions of the code which passed the acceptance suite
but were not idempotent, including critical failures such as causing
each run to attempt to uninstall and reinstall all instance features.

This commit updates the logic of the ensure_sqlserver_instance
helper function to execute the specified manifest twice; once to
test for failures when applying and a second time to ensure no
changes are attempted when the manifest is re-applied. It also
raises the test tier for tests which rely on this function to
the high tier until after the next release.