(MODULES-6356) Fixes a problem still remaining from MODULES-2904
As described in MODULES-2904, installation fails with an array of SQL Admins.
This was fixed in the use_discrete=true block (line 171-178. However the double
quotes also needed to be removed from the list of accounts for the
/SQLSYSADMINACCOUNTS switch parameter.  This appears to be due to the arguments
being interpretted literally, including the double quotes when doing account
lookups.  Without the fix, the only setting that works is
/SQLSYSADMINACCOUNTS="BUILTIN\Administrators", all other settings (both local
and domain accounts/groups) will fail, particularly those with whitespace.