(MAINT) Add unreleased section to CHANGELOG
Prior to this commit the windows team was not tracking
changes prior to release in the changelog. Per the
KeepAChangelog standard, using an unreleased section
provides value to the users and developers both.
Using this section will make release prep easier and
allow users who want to leverage the module from master
to do so with clear knowledge of what changes have
been made to the code since the last release.

This commit adds the unreleased section to the CHANGELOG
and adds notes for the changes.