(MODULES-7801) Fix Gem Dependencies
PDKified modules depend on puppet-module-posix-system-r2.3 and
puppet-module-win-system-r2.3 to bring in gems such as beaker-pe
which previously were explicitly specified in the Gemfile. Bringing in
those dependencies requires version 0.3.10 or greater, which is not
specified in the Gemfile. This has caused cases where a lower version
is brought into the bundle and testing breaks in CI and on the local
machines.