(MAINT) Fix broken Jenkins Pipeline
This change fixes a bug in the acceptance tests where the wrong set of
features was being verified for sqlserver 2016. Sqlserver 2016 will not
install Connectivity tools in the scenario being tested. This is due to
a change in that version of SQL Server, not a mal function in the module
code.