(MODULES-8606) Add SQL Logins Tasks
Add a set of tasks for ad hoc management of SQL Logins. These tasks
gather information about the current state of logins on an instance and
then will allow the user to set properties of a specific user on an
instance.

Scenarios for this might be a dba that usually has the SA login disabled
might want to enable it momentarily for a specific task, and then
disable it again.

It can also be used to ensure that the state of users, like a disabled
SA account are in the state a dba expects them to be in.