(MODULES-8677) Make Role Resource Titles Unique
Prior to this change, the titles of the `sqlserver_tsql` resources
generated by the `sqlserver::role` and `sqlserver::role::permissions`
defined types were not unique enough. The module was not able to manage
the two roles with the same name but on two instances running on a node
or two roles with the same name on two different databases on the same
instance.

This change ensures that the `sqlserver_tsql` resource names include the
names of the instances and databases being targeted so that titles are
unique enough to enable that capability.