(MODULES-8685) Fixed erb Template For Login Test
The login_exists.sql erb template generates a dynamic sql statement that
has a syntax error. The template completes without error, but when run
against the server the syntax error that results in the T-SQL statement
causes the defined type to believe that a login exists even if it does
not.

This change fixes the syntax error so that the sql statement runs as
expected. It also adds an acceptance test to ensure regressions are not
introduced.