(MODULES-8749) Make job_name param required.
Prior to this change, even though the job_name param is defined as
required by the PowerShell script that powers the task, the json file
that describes the task defines that parameter to Bolt as Optional.

This causes the task to hang if the task is invoked with no job_name
parameter. This change updates the json so that Bolt understands the
param is required, and removes an unnecessary conditional check to
to ensure the paramter is present, as both platforms already ensure
it. The else block in that check also had code that should never run
in this task.