(MAINT) Increase timeout interval
The wait parameter in the start_sql_agent_job depends on a sleep
interval that waits long enough for the agent to spool up a job and get
it running.

In testing it appears that some agent jobs are not spinning up fast
enough, causing the wait parameter to have no effect. This change
increased the timeout so that the wait parameter has time to see the
job starting to execute and then wait for it to complete.