(MODULES-7553) Add Example Clarification
A user trying to follow our advanced example ran into an issue. The
example as written is only suitable for setups that are using local
accounts, not domain accounts. They also found that the Facter fact
`domain` returns the FQDN of the domain, which is not suitable for
contsructing user names in the `netbiosname\username` format.

This change adds a clarification to the example that the example is only
usable when using local group, and given that there is no `netbiosname`
fact available, provides a suggestion for one way to work around the
issue.