(MODULES-6904) Restart SQL Server Service with Dependent Services
Prviously, the sp_configure.pp file would attempt to use the builtin
service resource to restart the MSSQLServer service if a configuration
change was made. This attempt would fail if there was a dependent
service such as the SQL Agent service, which is extremely common.

This change implements an extremely small PowerShell script instead to
force restart the MSSQLSERVER service, even if there are dependent
services.

In testing this resulted in the Agent service going down and then
coming back up as expected.