(MODULES-9912) ERB Template Errors
Prior to this change, some ERB templates would cause jruby errors
on the master during catalog compilation. The templates contained code
that would attempt to modify the value of a frozen string.

This change modifies the template code to ensure that no frozen string
values are attempted to be modified.

It also makes slight fixes to the syntax in the templates to ensure that
each instance of this pattern is written more consistantly.