(FM-8879) Handle T-SQL Errors Properly
The previous commit that improved T-SQL error handling was incomplete.
That commit (9b8a0fe9bbbe0699b6626f92ac6f9dc4ddad9584) was not able to
overcome an important issue.

The issue was that if there was a successfull row insert into a table
the `Errors` collection of the ADODB.Connection would be empty, even
if an error was deliberately thrown using the `THROW` keyword.

It turns out that the successfull row insert was causing a text return
value, and it's a known issue that if values are returned to the object
that the Errors collection will be empty.

We know now that turning `SET NOCOUNT ON` in the beginning of the query
will suppress that text status message and allow more errors to be
properly detected in the ADODB.Connect object, and handled properly
in the module's `onlyif` clause.