(CONT-370) Reinstate lost sqlserver documentation on feature & tool deprecations
Prior to this commit, using puppetlabs-sqlserver with deprecated features/tools would raise a deprecation warning messages in puppet agent runs, although the documentation says the deprecated variable is valid. This is due to a mistake in a puppet-strings run which excluded part of the documentation regarding this issue. This commit reinstate this documentation to avoid future confusion.