(Bug) - Remove default switch of updateenabled false
Prior to this commit, since v3.3.0 of this module,
the Install switch /UpdateEnabled always defaulted
to FALSE when the instance /ACTION=Install.

This took away the ability for users to be able to
configure this install switch themselves,
and is an obvious bug.

Now, this no longer automatically prepends the flag to
install switches when action=install, and allows users
to configure this through install_switches property
once again.